requirements.txt
absl-py==1.0.0
anyascii==0.3.0
astunparse==1.6.3
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
contractions==0.0.55
falcon==3.0.1
filelock==3.3.2
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.5.0
huggingface-hub==0.1.2
hupper==1.10.3
idna==3.3
importlib-metadata==4.8.2
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.4
nltk==3.6.5
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.2
pandas==1.3.4
Pillow==8.4.0
protobuf==3.19.1
pyahocorasick==1.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
textsearch==0.0.21
tokenizers==0.10.3
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
transformers==4.12.3
typing-extensions==3.10.0.2
urllib3==1.26.7
waitress==2.0.0
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
